feat(services): use a single service liveness coordinator
