+++ core/src/main/java/io/kestra/core/server/DefaultServiceLivenessCoordinator.java
@@ -1,5 +1,9 @@
 package io.kestra.core.server;
 
+import com.google.common.annotations.VisibleForTesting;
+import io.kestra.core.lock.Lock;
+import io.kestra.core.lock.LockService;
+import io.kestra.core.metrics.MetricRegistry;
 import io.kestra.core.queues.QueueException;
 import io.kestra.core.queues.QueueFactoryInterface;
 import io.kestra.core.queues.QueueInterface;
@@ -7,17 +11,19 @@ import io.kestra.core.repositories.ServiceInstanceRepositoryInterface;
 import io.kestra.core.runners.*;
 import io.kestra.core.services.LogService;
 import io.kestra.core.services.SkipExecutionService;
-import io.micronaut.core.annotation.Introspected;
+import io.kestra.core.utils.IdUtils;
+import io.micronaut.context.annotation.Requires;
+import io.micronaut.context.annotation.Value;
+import io.micronaut.scheduling.annotation.Scheduled;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
+import jakarta.inject.Singleton;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.event.Level;
 
 import java.time.Duration;
 import java.time.Instant;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Random;
 import java.util.Set;
@@ -25,57 +31,79 @@ import java.util.Set;
 import static io.kestra.core.server.Service.ServiceState.*;
 
 /**
- * Base class for coordinating service liveness.
+ * Responsible for coordinating the state of all service instances.
+ *
+ * @see ServiceInstance
  */
-@Introspected
 @Slf4j
-public abstract class AbstractServiceLivenessCoordinator extends AbstractServiceLivenessTask {
+@Singleton
+@Requires(property = "kestra.server-type", pattern = "(EXECUTOR|STANDALONE)")
+public class DefaultServiceLivenessCoordinator extends AbstractServiceLivenessTask {
 
-    private final static int DEFAULT_SCHEDULE_JITTER_MAX_MS = 500;
+     private static final int DEFAULT_SCHEDULE_JITTER_MAX_MS = 500;
 
-    protected static String DEFAULT_REASON_FOR_DISCONNECTED =
+    private static final String DEFAULT_REASON_FOR_DISCONNECTED =
         "The service was detected as non-responsive after the session timeout. " +
         "Service transitioned to the 'DISCONNECTED' state.";
 
-    protected static String DEFAULT_REASON_FOR_NOT_RUNNING =
+    private static final String DEFAULT_REASON_FOR_NOT_RUNNING =
         "The service was detected as non-responsive or terminated after termination grace period. " +
         "Service transitioned to the 'NOT_RUNNING' state.";
 
     private static final String TASK_NAME = "service-liveness-coordinator-task";
 
-    protected final ServiceLivenessStore store;
+    private final ServiceLivenessStore store;
+    private final ServiceRegistry serviceRegistry;
+    private final ServiceLivenessUpdater serviceLivenessUpdater;
+    private final ServiceInstanceRepositoryInterface serviceInstanceRepository;
+    private final Duration purgeRetention;
 
-    protected final ServiceRegistry serviceRegistry;
-
+    private final LockService lockService;
     private final SkipExecutionService skipExecutionService;
     private final QueueInterface<WorkerJob> workerJobQueue;
     private final WorkerJobRunningStateStore workerJobRunningStateStore;
     private final LogService logService;
+    private final MetricRegistry metricRegistry;
 
     // mutable for testing purpose
-    protected String serverId = ServerInstance.INSTANCE_ID;
+    String serverId = ServerInstance.INSTANCE_ID;
+
+    @VisibleForTesting
+    void setServerInstance(String serverId) {
+        this.serverId = serverId;
+    }
 
     /**
-     * Creates a new {@link AbstractServiceLivenessCoordinator} instance.
+     * Creates a new {@link DefaultServiceLivenessCoordinator} instance.
      *
      * @param store        The {@link ServiceInstanceRepositoryInterface}.
      * @param serverConfig The server configuration.
      */
     @Inject
-    public AbstractServiceLivenessCoordinator(final ServiceLivenessStore store,
-                                              final ServiceRegistry serviceRegistry,
-                                              final SkipExecutionService skipExecutionService,
-                                              final @Named(QueueFactoryInterface.WORKERJOB_NAMED) QueueInterface<WorkerJob> workerJobQueue,
-                                              final WorkerJobRunningStateStore workerJobRunningStateStore,
-                                              final LogService logService,
-                                              final ServerConfig serverConfig) {
+    public DefaultServiceLivenessCoordinator(final ServiceLivenessStore store,
+                                             final ServiceRegistry serviceRegistry,
+                                             final ServiceLivenessUpdater serviceLivenessUpdater,
+                                             final ServiceInstanceRepositoryInterface serviceInstanceRepository,
+                                             final LockService lockService,
+                                             final SkipExecutionService skipExecutionService,
+                                             final @Named(QueueFactoryInterface.WORKERJOB_NAMED) QueueInterface<WorkerJob> workerJobQueue,
+                                             final WorkerJobRunningStateStore workerJobRunningStateStore,
+                                             final LogService logService,
+                                             final ServerConfig serverConfig,
+                                             final MetricRegistry metricRegistry,
+                                             @Value("${kestra.server.service.purge.retention}") final Duration purgeRetention) {
         super(TASK_NAME, serverConfig);
         this.serviceRegistry = serviceRegistry;
+        this.serviceLivenessUpdater = serviceLivenessUpdater;
+        this.serviceInstanceRepository = serviceInstanceRepository;
         this.store = store;
         this.skipExecutionService = skipExecutionService;
         this.workerJobQueue = workerJobQueue;
         this.workerJobRunningStateStore = workerJobRunningStateStore;
         this.logService = logService;
+        this.lockService = lockService;
+        this.metricRegistry = metricRegistry;
+        this.purgeRetention = purgeRetention;
     }
 
     /**
@@ -108,26 +136,90 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
     }
 
     /**
-     * Handles all unresponsive services and update their status to disconnected.
-     * <p>
-     * This method may re-submit tasks is necessary.
-     *
-     * @param now the time of the execution.
-     */
-    protected abstract void handleAllNonRespondingServices(final Instant now);
-
-    /**
      * Handles all worker services which are shutdown or considered to be terminated.
      * <p>
      * This method may re-submit tasks is necessary.
      *
      * @param now the time of the execution.
      */
-    protected abstract void handleAllWorkersForUncleanShutdown(final Instant now);
+    protected void handleAllWorkersForUncleanShutdown(Instant now) {
+            serviceInstanceRepository.processAllNonRunningInstances((txContext, serviceInstance) -> {
+                // List of workers for which we don't know the actual state of tasks executions.
+                // Re-emit all WorkerJobs for unclean workers
+                if (isUncleanShutdownService(serviceInstance, now)) {
+                    if (serviceInstance.config().workerTaskRestartStrategy().isRestartable()) {
+                        log.info("Trigger task restart for non-responding worker after termination grace period: {}.", serviceInstance.uid());
+                        reEmitWorkerJobsForWorker(txContext, serviceInstance.uid());
+                    }
+                }
 
-    protected abstract void update(final ServiceInstance instance,
-                                   final Service.ServiceState state,
-                                   final String reason) ;
+                // Transit GRACEFUL and UNCLEAN SHUTDOWN worker to NOT_RUNNING.
+                if (serviceInstance.is(Service.ServiceState.TERMINATED_GRACEFULLY)) {
+                    serviceInstanceRepository.mayTransitServiceTo(txContext,
+                        serviceInstance,
+                        Service.ServiceState.NOT_RUNNING,
+                        DEFAULT_REASON_FOR_NOT_RUNNING
+                    );
+                }
+        });
+    }
+
+    /**
+     * {@inheritDoc}
+     **/
+    protected void update(ServiceInstance instance, Service.ServiceState state, String reason) {
+        serviceLivenessUpdater.update(instance, state, reason);
+    }
+
+    /**
+     * Handles all unresponsive services and update their status to disconnected.
+     * <p>
+     * This method may re-submit tasks is necessary.
+     *
+     * @param now the time of the execution.
+     */
+    protected void handleAllNonRespondingServices(Instant now) {
+        // Retrieves all services that are supposed to be running.
+        serviceInstanceRepository.processInstanceInStates(allRunningStates(), (txContext, serviceInstance) -> {
+            // Detect and handle non-responding services.
+            if (isNonRespondingService(serviceInstance, now)) {
+                // Attempt to transit all non-responding services to DISCONNECTED.
+                serviceInstanceRepository.mayTransitServiceTo(
+                    txContext,
+                    serviceInstance,
+                    Service.ServiceState.DISCONNECTED,
+                    DEFAULT_REASON_FOR_DISCONNECTED
+                );
+
+                // Eventually restart worker tasks
+                if (serviceInstance.is(ServiceType.WORKER) &&
+                    serviceInstance.config().workerTaskRestartStrategy().equals(WorkerTaskRestartStrategy.IMMEDIATELY)) {
+                    log.info("Trigger task restart for non-responding worker after timeout: {}.", serviceInstance.uid());
+                    reEmitWorkerJobsForWorker(txContext, serviceInstance.uid());
+                }
+
+                // Eventually release all owned locks
+                List<Lock> released = lockService.releaseAllLocks(serviceInstance.uid());
+                released.forEach(l -> log.info("Released lock {} for non-responding service instance {}", IdUtils.fromParts(l.getCategory(), l.getId()), serviceInstance.uid()));
+            }
+        });
+    }
+
+    @Scheduled(initialDelay = "${kestra.server.service.purge.initial-delay}", fixedDelay = "${kestra.server.service.purge.fixed-delay}")
+    public void purgeEmptyInstances() {
+        int purged = serviceInstanceRepository.purgeEmptyInstances(Instant.now().minus(purgeRetention));
+        log.info("Purged {} service instances", purged);
+    }
+
+    private void reEmitWorkerJobsForWorker(final TransactionContext txContext, final String id) {
+        metricRegistry.counter(MetricRegistry.METRIC_EXECUTOR_WORKER_JOB_RESUBMIT_COUNT, MetricRegistry.METRIC_EXECUTOR_WORKER_JOB_RESUBMIT_COUNT_DESCRIPTION)
+            .increment();
+
+        workerJobRunningStateStore.processWorkerJobsForDeadWorker(txContext, id, (txContext2, workerJobRunning) -> {
+            resubmitWorkerJobRunning(txContext2, workerJobRunning);
+        });
+    }
+
 
     /**
      * {@inheritDoc}
@@ -142,60 +234,49 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
         return serverConfig.liveness().interval().plus(Duration.ofMillis(jitter));
     }
 
-    protected List<ServiceInstance> filterAllUncleanShutdownServices(final List<ServiceInstance> instances,
-                                                                     final Instant now) {
-        // List of services for which we don't know the actual state
-        final List<ServiceInstance> uncleanShutdownServices = new ArrayList<>();
-
+    private boolean isUncleanShutdownService(final ServiceInstance instance, final Instant now) {
         // ...all services that have transitioned to DISCONNECTED or TERMINATING for more than terminationGracePeriod.
-        uncleanShutdownServices.addAll(instances.stream()
-            .filter(nonRunning -> nonRunning.state().isDisconnectedOrTerminating())
-            .filter(disconnectedOrTerminating -> disconnectedOrTerminating.isTerminationGracePeriodElapsed(now))
-            .peek(instance -> maybeLogNonRespondingAfterTerminationGracePeriod(instance, now))
-            .toList()
-        );
+        if (instance.state().isDisconnectedOrTerminating() && instance.isTerminationGracePeriodElapsed(now)) {
+            maybeLogNonRespondingAfterTerminationGracePeriod(instance, now);
+            return true;
+        }
+
         // ...all services that have transitioned to TERMINATED_FORCED.
-        uncleanShutdownServices.addAll(instances.stream()
-            .filter(nonRunning -> nonRunning.is(Service.ServiceState.TERMINATED_FORCED))
-            // Only select workers that have been terminated for at least the grace period, to ensure that all in-flight
-            // task runs had enough time to be fully handled by the executors.
-            .filter(terminated -> terminated.isTerminationGracePeriodElapsed(now))
-            .toList()
-        );
-        return uncleanShutdownServices;
+        // Only select workers that have been terminated for at least the grace period, to ensure that all in-flight
+        // task runs had enough time to be fully handled by the executors.
+        return instance.is(Service.ServiceState.TERMINATED_FORCED) && instance.isTerminationGracePeriodElapsed(now);
     }
 
-    protected List<ServiceInstance> filterAllNonRespondingServices(final List<ServiceInstance> instances,
-                                                                   final Instant now) {
-        return instances.stream()
-            .filter(instance -> Objects.nonNull(instance.config())) // protect against non-complete instance
-            .filter(instance -> instance.config().liveness().enabled())
-            .filter(instance -> instance.isSessionTimeoutElapsed(now))
+    private boolean isNonRespondingService(final ServiceInstance instance, final Instant now) {
+        boolean isNonResponding = instance.config() != null && // protect against non-complete instance
+            instance.config().liveness().enabled() &&
+            instance.isSessionTimeoutElapsed(now) &&
             // exclude any service running on the same server as the executor, to prevent the latter from shutting down.
-            .filter(instance -> !instance.server().id().equals(serverId))
+            !instance.server().id().equals(serverId) &&
             // only keep services eligible for liveness probe
-            .filter(instance -> {
-                final Instant minInstantForLivenessProbe = now.minus(instance.config().liveness().initialDelay());
-                return instance.createdAt().isBefore(minInstantForLivenessProbe);
-            })
+            instance.createdAt().isBefore(now.minus(instance.config().liveness().initialDelay()));
+
             // warn
-            .peek(instance -> log.warn("Detected non-responding service [id={}, type={}, hostname={}] after timeout ({}ms).",
-                instance.uid(),
-                instance.type(),
-                instance.server().hostname(),
-                now.toEpochMilli() - instance.updatedAt().toEpochMilli()
-            ))
-            .toList();
+            if (isNonResponding) {
+                log.warn("Detected non-responding service [id={}, type={}, hostname={}] after timeout ({}ms).",
+                    instance.uid(),
+                    instance.type(),
+                    instance.server().hostname(),
+                    now.toEpochMilli() - instance.updatedAt().toEpochMilli()
+                );
+            }
+
+            return isNonResponding;
 
     }
 
-    protected void handleAllServiceInNotRunningState() {
+    private void handleAllServiceInNotRunningState() {
         // Soft delete all services which are NOT_RUNNING anymore.
         store.findAllInstancesInStates(Set.of(Service.ServiceState.NOT_RUNNING))
             .forEach(instance -> safelyUpdate(instance, Service.ServiceState.INACTIVE, null));
     }
 
-    protected void handleAllServicesForTerminatedStates(final Instant now) {
+    private void handleAllServicesForTerminatedStates(final Instant now) {
         store
             .findAllInstancesInStates(Set.of(DISCONNECTED, TERMINATING, TERMINATED_GRACEFULLY, TERMINATED_FORCED))
             .stream()
@@ -205,7 +286,7 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
             .forEach(instance -> safelyUpdate(instance, NOT_RUNNING, DEFAULT_REASON_FOR_NOT_RUNNING));
     }
 
-    protected void maybeDetectAndLogNewConnectedServices() {
+    private void maybeDetectAndLogNewConnectedServices() {
         if (log.isDebugEnabled()) {
             // Log the newly-connected services (useful for troubleshooting).
             store.findAllInstancesInStates(Set.of(CREATED, RUNNING))
@@ -222,7 +303,7 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
         }
     }
 
-    protected void safelyUpdate(final ServiceInstance instance,
+    private void safelyUpdate(final ServiceInstance instance,
                                 final Service.ServiceState state,
                                 final String reason) {
         try {
@@ -240,7 +321,7 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
         }
     }
 
-    protected void maybeLogNonRespondingAfterTerminationGracePeriod(final ServiceInstance instance,
+    private void maybeLogNonRespondingAfterTerminationGracePeriod(final ServiceInstance instance,
                                                                            final Instant now) {
         if (instance.state().isDisconnectedOrTerminating()) {
             log.warn("Detected non-responding service [id={}, type={}, hostname={}] after termination grace period ({}ms).",
@@ -252,7 +333,7 @@ public abstract class AbstractServiceLivenessCoordinator extends AbstractService
         }
     }
 
-    protected void resubmitWorkerJobRunning(TransactionContext txContext, WorkerJobRunning workerJobRunning) {
+    private void resubmitWorkerJobRunning(TransactionContext txContext, WorkerJobRunning workerJobRunning) {
         // WorkerTaskRunning
         if (workerJobRunning instanceof WorkerTaskRunning workerTaskRunning) {
             if (skipExecutionService.skipExecution(workerTaskRunning.getTaskRun())) {
