+++ core/src/main/java/io/kestra/core/server/ServiceLivenessStore.java
@@ -10,16 +10,25 @@ import java.util.Set;
  *
  * @see ServiceInstance
  * @see ServiceLivenessUpdater
- * @see AbstractServiceLivenessCoordinator
+ * @see DefaultServiceLivenessCoordinator
  */
 public interface ServiceLivenessStore  {
 
     /**
-     * Finds all service instances which are in one of the given states.
+     * Finds all service instances that are in one of the given states.
      *
      * @param states the state of services.
      *
      * @return the list of {@link ServiceInstance}.
      */
     List<ServiceInstance> findAllInstancesInStates(Set<ServiceState> states);
+
+    /**
+     * Finds all service instances that in the given state.
+     *
+     * @param state the state of services.
+     *
+     * @return the list of {@link ServiceInstance}.
+     */
+    List<ServiceInstance> findAllInstancesInState(ServiceState state);
 }
